chore: move utils out of songs
