+++ billboard/songs/hot.py
@@ -1,8 +1,8 @@
 from datetime import datetime, timedelta
 from typing import Optional
 
+from ..utils import make_request, parse_request
 from .super import SongChart
-from .utils import make_request, parse_request
 
 
 class BillboardChart(SongChart):
